Create own shared key if no key was obtained from peers
This extends the logic to not just creating a shared key if nobody was found in the lobby,
but also creating one if nodes are found but they are not responding with a shared key.

This is a bit dangerous as it creates a situation one can't *easily* get out of,
but it is needed if two nodes boot up and none of the two has a shared key yet.

There are obvious race conditions, which would ideally be resolved by waiting
a random time, or some other protocol, so that even if two nodes are restarted
at precisely the same moment, no split state is created.